+++ schema/execute_msg.json
@@ -11,11 +11,19 @@
         "approve_ask": {
           "type": "object",
           "required": [
-            "id"
+            "base",
+            "id",
+            "size"
           ],
           "properties": {
+            "base": {
+              "type": "string"
+            },
             "id": {
               "type": "string"
+            },
+            "size": {
+              "$ref": "#/definitions/Uint128"
             }
           }
         }
